requirements.txt
fastapi
uvicorn
openai
faiss-cpu
numpy
python-dotenv
